// frontend/src/api/api.ts
export interface DtosLoginRequest {
    /**
     * 
     * @type {string}
     * @memberof DtosLoginRequest
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof DtosLoginRequest
     */
    'username'?: string;
}
// ...
export interface DtosRegisterRequest {
    /**
     * 
     * @type {string}
     * @memberof DtosRegisterRequest
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof DtosRegisterRequest
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof DtosRegisterRequest
     */
    'username'?: string;
}
// ...
export class AuthApi extends BaseAPI {
    /**
     * Authenticates a user and generates a token upon successful login
     * @summary Login
     * @param {DtosLoginRequest} loginRequest Login request payload
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authLoginPost(loginRequest: DtosLoginRequest, options?: RawAxiosRequestConfig) {
        return AuthApiFp(this.configuration).authLoginPost(loginRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Creates a new user account and returns authentication token
     * @summary Register a new user
     * @param {DtosRegisterRequest} request Registration Details
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authRegisterPost(request: DtosRegisterRequest, options?: RawAxiosRequestConfig) {
        return AuthApiFp(this.configuration).authRegisterPost(request, options).then((request) => request(this.axios, this.basePath));
    }
}